requirements.txt
hwi
flask
python-bitcoinrpc

flask_json_schema
# with simplejson flask can serialize decimals
simplejson
flask-cors

# wallets
ckcc-protocol
trezor

# pyinstaller
cefpython3
libusb1
pywebview[qt]
pywebview[cef]
pyinstaller.zip

